+++ magnet/example_random_instance_predictor.py
@@ -1,25 +1,25 @@
 import random
 import argparse
 
-from magnet.predictor import RunPredictor, RunPrediction
+from magnet.instance_predictor import InstancePredictor, InstancePrediction
 from magnet.data_splits import TrainSplit, SequesteredTestSplit
 
 
-class ExampleRandomPredictor(RunPredictor):
+class ExampleRandomInstancePredictor(InstancePredictor):
     """
-    Class to demonstrate a random stat prediction algorithm
+    Class to demonstrate a random per-instance stat prediction algorithm
 
     Example:
         >>> import magnet
         >>> outputs = magnet.HelmOutputs.demo()
         >>> suite_path = outputs.suites()[0].path
-        >>> predictor_instance = ExampleRandomPredictor(num_eval_samples=5)
+        >>> predictor_instance = ExampleRandomInstancePredictor(num_eval_samples=5)
         >>> predictor_instance(suite_path)
     """
     def predict(self,
                 train_split: TrainSplit,
                 sequestered_test_split: SequesteredTestSplit
-                ) -> list[RunPrediction]:
+                ) -> list[InstancePrediction]:
         # Unpack split classes into dataframes
         train_run_specs_df = train_split.run_specs  # NOQA
         train_scenario_states_df = train_split.scenario_state  # NOQA
@@ -29,15 +29,16 @@ class ExampleRandomPredictor(RunPredictor):
         eval_scenario_state_df = sequestered_test_split.scenario_state
 
         predictions = []
+        for _, row in eval_scenario_state_df.iterrows():
+            run_spec_name = row['run_spec.name']
+            instance_predict_id = row['magnet.instance_predict_id']
 
-        for key, _ in eval_scenario_state_df.groupby(['run_spec.name']):
-            run_spec_name, = key
-            prediction = (random.choice(range(0, 101)) / 100)
+            prediction = random.choice([0.0, 1.0])
 
             predictions.append(
-                RunPrediction(
+                InstancePrediction(
                     run_spec_name=run_spec_name,
-                    split="valid",
+                    instance_predict_id=instance_predict_id,
                     stat_name="exact_match",
                     mean=prediction))
 
@@ -45,7 +46,7 @@ class ExampleRandomPredictor(RunPredictor):
 
 def main():
     parser = argparse.ArgumentParser(
-        description="Run example random predictor")
+        description="Run example random per-instance predictor")
 
     parser.add_argument('helm_suite_path',
                         type=str,
@@ -53,7 +54,7 @@ def main():
 
     args = parser.parse_args()
 
-    predictor_instance = ExampleRandomPredictor()
+    predictor_instance = ExampleRandomInstancePredictor()
     predictor_instance(args.helm_suite_path)